[feature] move relations to concern trait for model `Project`
